+++ src/main/java/com/anz/account/repository/AccountDetailsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountDetailsRepository extends JpaRepository<Account, Integer> {
+public interface AccountDetailsRepository extends JpaRepository<Account, Long> {
 
 
 }
